Fix tests and reorganize (#100)
