+++ tests/unit/test_base_chat_session.py
@@ -142,3 +142,27 @@ def test_new_and_import_state():
     assert tgt.session_id is None and tgt.session_alias is None
     assert tgt.session_title is None
     assert tgt.history.get_messages() == []
+
+
+class PassThroughSession(BaseChatSession):
+    def __init__(self):
+        super().__init__()
+
+    def invoke(self, messages=None, disable_system_prompt=False, model=None, temperature=None):
+        super().invoke(messages, disable_system_prompt)
+        return "ok"
+
+    def invoke_with_tools(self, messages=None, disable_system_prompt=False):
+        super().invoke_with_tools(messages, disable_system_prompt)
+        return "ok", []
+
+    def list_models(self, ignore_errors=False):
+        super().list_models(ignore_errors=ignore_errors)
+        return []
+
+
+def test_super_methods_execute():
+    session = PassThroughSession()
+    assert session.invoke() == "ok"
+    assert session.invoke_with_tools() == ("ok", [])
+    assert session.list_models() == []
