+++ tests/unit/test_chat_commands.py
@@ -1,6 +1,6 @@
 import argparse
 import pytest
-from tests.test_chat_interface_extended import make_interface
+from tests.unit.test_chat_interface_extended import make_interface
 import lair
