+++ tests/unit/test_chat_interface_commands.py
