+++ tests/unit/test_chat_interface_extended.py
