+++ tests/unit/test_chat_interface_reports.py
