+++ tests/unit/test_config_module.py
@@ -61,3 +61,25 @@ def test_update_get_set_and_cast(monkeypatch, tmp_path):
         cfg.set("chat.attachments_enabled", "maybe")
     # cast empty string for int returns None
     assert cfg._cast_value("session.max_history_length", "") is None
+
+
+def test_add_config_inherit_and_no_default(tmp_path, monkeypatch):
+    make_config_env(tmp_path, monkeypatch)
+    cfg = config_module.Configuration()
+    new_config = {
+        "base": {"chat.enable_toolbar": False},
+        "child": {"_inherit": ["base"], "chat.attachments_enabled": True},
+    }
+    cfg._add_config(new_config)
+    assert cfg.modes["child"]["chat.enable_toolbar"] is False
+    assert cfg.modes["child"]["chat.attachments_enabled"] is True
+    assert cfg.active_mode == "test"
+
+
+def test_set_inherit_and_invalid_cast(tmp_path, monkeypatch):
+    make_config_env(tmp_path, monkeypatch)
+    cfg = config_module.Configuration()
+    cfg.set("_inherit", ["base"])
+    assert cfg.get("_inherit") == ["base"]
+    with pytest.raises(config_module.ConfigInvalidType):
+        cfg.set("session.max_history_length", "oops")
