+++ tests/unit/test_logging_and_loader.py
