+++ tests/unit/test_openai_chat_session.py
