+++ tests/unit/test_sessions_init.py
